refactor: 필터링 Component props 적용
