+++ packages/client/src/components/contentPanel/filter/CreditFilter.tsx
@@ -1,4 +1,5 @@
 import { useScheduleSearchStore } from '@/store/useFilterStore.ts';
+import CheckboxAdapter from '@common/components/checkbox/CheckboxAdapter';
 import MultiCheckboxFilter from '@common/components/filtering/MultiCheckboxFilter';
 import { OptionType } from '@common/components/filtering/MultiCheckboxFilter';
 
@@ -26,6 +27,7 @@ function CreditFilter() {
       setFilterSchedule={setFilterScheduleWrapper}
       options={CREDITS}
       selected={credits.length !== 0}
+      ItemComponent={CheckboxAdapter}
     />
   );
 }
