+++ packages/client/src/components/simulation/modal/before/TutorialModal.tsx
@@ -5,7 +5,7 @@ import { useSimulationModalStore } from '@/store/simulation/useSimulationModal';
 import ArrowdownSvg from '@/assets/arrow-down-gray.svg?react';
 import YouTube from 'react-youtube';
 import useMobile from '@/hooks/useMobile';
-import Checkbox from '@common/components/filtering/Checkbox';
+import Checkbox from '@common/components/checkbox/Checkbox';
 
 const tutorialVideos = [
   {
