+++ packages/common/src/components/filtering/MultiCheckboxFilter.tsx
@@ -1,5 +1,3 @@
-import Chip from '../chip/Chip';
-import Checkbox from './Checkbox';
 import Filtering from './Filtering';
 
 export interface OptionType<VALUE extends string | number> {
@@ -7,28 +5,33 @@ export interface OptionType<VALUE extends string | number> {
   label: string;
 }
 
-type Variant = 'checkbox' | 'chip';
+interface FilterItemProps<VALUE extends string | number> {
+  label: string;
+  selected: boolean;
+  onClick: () => void;
+  value: VALUE;
+}
 
 interface IMultiCheckboxFilter<VALUE extends string | number> {
   labelPrefix: string;
   selectedValues: VALUE[];
   field: string;
-  variant?: Variant;
   setFilterSchedule: (field: string, value: VALUE[]) => void;
   options: OptionType<VALUE>[];
-  selected: boolean;
   className?: string;
+  selected: boolean;
+  ItemComponent: React.ComponentType<FilterItemProps<VALUE>>;
 }
 
 function MultiCheckboxFilter<VALUE extends string | number>({
   labelPrefix,
   selectedValues,
   field,
-  variant = 'checkbox',
   setFilterSchedule,
   options,
-  selected,
   className = '',
+  selected,
+  ItemComponent,
 }: Readonly<IMultiCheckboxFilter<VALUE>>) {
   const checkSelected = (value: VALUE) => {
     return selectedValues.includes(value);
@@ -64,25 +67,15 @@ function MultiCheckboxFilter<VALUE extends string | number>({
     <Filtering label={getFilteringLabel()} selected={selected} className={className}>
       <h3 className="text-md font-medium text-gray-600">{labelPrefix}</h3>
       <div className={`${options.length > 6 ? 'grid grid-cols-2 gap-x-4' : 'flex flex-col'} gap-y-2`}>
-        {variant === 'checkbox' &&
-          options.map(option => (
-            <Checkbox
-              key={String(option.value)}
-              label={option.label}
-              checked={checkSelected(option.value)}
-              onChange={() => handleChangeCheckbox(option.value)}
-            />
-          ))}
-
-        {variant === 'chip' &&
-          options.map(option => (
-            <Chip
-              key={String(option.value)}
-              label={option.label}
-              selected={checkSelected(option.value)}
-              onClick={() => handleChangeCheckbox(option.value)}
-            />
-          ))}
+        {options.map(option => (
+          <ItemComponent
+            key={String(option.value)}
+            label={option.label}
+            selected={checkSelected(option.value)}
+            onClick={() => handleChangeCheckbox(option.value)}
+            value={option.value}
+          />
+        ))}
       </div>
       <div className="flex justify-end w-full mt-2">
         <button onClick={() => handleClickReset()} className="text-blue-500 cursor-pointer text-sm px-1 py-0.5">
