+++ packages/common/src/components/filtering/SingleCheckbox.tsx
@@ -1,5 +1,3 @@
-import Chip from '../chip/Chip';
-import Checkbox from './Checkbox';
 import Filtering from './Filtering';
 
 export interface OptionType<VALUE extends string | number> {
@@ -7,28 +5,33 @@ export interface OptionType<VALUE extends string | number> {
   label: string;
 }
 
-type Variant = 'checkbox' | 'chip';
+interface FilterItemProps<VALUE extends string | number> {
+  label: string;
+  selected: boolean;
+  onClick: () => void;
+  value: VALUE;
+}
 
 interface ISingleCheckboxFilter<VALUE extends string | number> {
   labelPrefix: string;
   selectedValue: VALUE | null;
   field: string;
-  variant?: Variant;
   setFilterSchedule: (field: string, value: VALUE | null) => void;
   options: OptionType<VALUE>[];
   selected: boolean;
   className?: string;
+  ItemComponent: React.ComponentType<FilterItemProps<VALUE>>;
 }
 
 function SingleCheckboxFilter<VALUE extends string | number>({
   labelPrefix,
   selectedValue,
   field,
-  variant = 'checkbox',
   setFilterSchedule,
   options,
   selected,
   className = '',
+  ItemComponent,
 }: Readonly<ISingleCheckboxFilter<VALUE>>) {
   const getFilteringLabel = () => {
     if (!selectedValue) {
@@ -53,26 +56,16 @@ function SingleCheckboxFilter<VALUE extends string | number>({
     <Filtering label={getFilteringLabel()} selected={selected} className={className}>
       <h3 className="text-gray-700 font-semibold">{labelPrefix}</h3>
 
-      {variant === 'checkbox' &&
-        options.map(option => (
-          <Checkbox
+      <div className="grid grid-cols-2 gap-2">
+        {options.map(option => (
+          <ItemComponent
             key={String(option.value)}
             label={option.label}
-            checked={selectedValue === option.value}
-            onChange={() => handleChangeCheckbox(option.value)}
+            selected={selectedValue === option.value}
+            onClick={() => handleChangeCheckbox(option.value)}
+            value={option.value}
           />
         ))}
-
-      <div className="grid grid-cols-2 gap-2">
-        {variant === 'chip' &&
-          options.map(option => (
-            <Chip
-              key={String(option.value)}
-              label={option.label}
-              selected={selectedValue === option.value}
-              onClick={() => handleChangeCheckbox(option.value)}
-            />
-          ))}
       </div>
 
       <div className="flex justify-end w-full mt-2">
